add static method to gitignore
